requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
twilio==5.4.0
django-twilio==0.8.0
django-sendsms==0.2.3
